% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.ftime.R
\name{bin_ftime}
\alias{bin_ftime}
\title{bin_ftime}
\usage{
bin_ftime(ftime, bins = "default", nKnots = 10)
}
\arguments{
\item{ftime}{A vector of failure times}

\item{bins}{How should the failure times be discretized? Can be set to 
\code{"default"} in which case cut points will be evenly over the range of \code{ftime}. 
Or can be set to a vector containing the desired cutpoints for the failure time bins.}

\item{nKnots}{If \code{bins="default"}, how many bins of failure times should there be?
The default value is 10 and the option is ignored if \code{bins != "default"}}
}
\value{
A vector of binned failure times
}
\description{
This is a helper function that takes a vector of continuously valued failure times and
discretizes into \code{nKnots} number of discrete times. By default the bins
will be evenly spaced, but users can also input their own cutpoints through the
\code{bins} argument.
}
\examples{
ftime <- runif(100,0,10)
# first do default binning
binned.ftime <- bin_ftime(ftime = ftime)
table(binned.ftime)
# now specify own cutpoints
binned.ftime2 <- bin_ftime(ftime = ftime, bins = c(2,5,8))
table(binned.ftime)
}
