% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle-confint-method.R
\name{confint.survtmle}
\alias{confint.survtmle}
\title{confint.survtmle}
\usage{
\method{confint}{survtmle}(object, parm = 1:length(object$est),
  level = 0.95)
}
\arguments{
\item{object}{An object of class \code{survtmle}.}

\item{parm}{A numeric vector indicating which indices of \code{object$est} to
return confidence intervals for (default is to return all).}

\item{level}{The confidence level requested.}
}
\value{
A matrix with columns giving the lower and upper confidence limits
for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % 
(by default 2.5% and 97.5%).
}
\description{
Computes confidence intervals for a fitted \code{survtmle} object.
}
\examples{
# simulate data
set.seed(1234)
n <- 100
ftime <- round(runif(n,1,4))
ftype <- round(runif(n,0,2))
trt <- rbinom(n,1,0.5) 
adjustVars <- data.frame(W1 = rnorm(n), W2 = rnorm(n))

# Fit 1
# fit a survtmle object 
fit1 <- survtmle(ftime = ftime, ftype = ftype, trt = trt, adjustVars = adjustVars,
glm.trt = "W1 + W2", 
glm.ftime = "trt + W1 + W2", glm.ctime = "trt + W1 + W2", 
method="mean", t0=4)
# get confidence intervals
ci <- confint(fit1)
# ci
}
