% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_survtmle_internal.R
\name{cv_survtmle_internal}
\alias{cv_survtmle_internal}
\title{cv_survtmle_internal}
\usage{
cv_survtmle_internal(ftime, ftype, trt, adjustVars, t0 = max(ftime),
  incidence = TRUE, survtmleFits, validFolds, returnIC = TRUE,
  returnModels = TRUE, ftypeOfInterest = "all", trtOfInterest = "all",
  method = "hazard", bounds = NULL, verbose = TRUE,
  tol = 1/(length(ftime)), maxIter = 100)
}
\arguments{
\item{ftime}{A numeric vector of failure times. Right-censored observations should have corresponding 
\code{ftype} set to 0.}

\item{ftype}{A numeric vector indicating the type of failure. Observations with \code{ftype=0} are treated 
as a right-censored observation. Each unique value besides zero is treated as a separate
type of failure.}

\item{trt}{A numeric vector indicating observed treatment assignment. Each unique value will 
be treated as an (unordered) separate type of treatment. Currently, only two unique values of 
\code{trt} are supported.}

\item{adjustVars}{A data.frame of adjustment variables that will be used in estimating the 
conditional treatment, censoring, and failure (hazard or conditional mean) probabilities.}

\item{t0}{The time at which to return cumulative incidence estimates. By default this is set
to \code{max(ftime)}.}

\item{incidence}{If \code{TRUE} the function return estimates of cumulative incidence. If 
\code{FALSE} the function returns esimtates of survival.}

\item{survtmleFits}{A list of objects of class \code{survtmle} fit on train samples of data}

\item{validFolds}{A \code{list} of \code{vectors} indicating which observations were in the 
validation fold for each of the \code{survtmleFits}.}

\item{returnIC}{A boolean indicating whether to return vectors of influence curve estimates. These are
needed for some post-hoc comparisons, so it is recommended to leave as \code{TRUE} (the default)
unless the user is sure these estimates will not be needed later.}

\item{returnModels}{A boolean indicating whether to return the \code{SuperLearner} or \code{glm} 
objects used to estimate the nuisance parameters. Must be set to \code{TRUE} if the user plans to 
use calls to \code{timepoints} to obtain estimates at times other than \code{t0}. See \code{?timepoints}
for more information.}

\item{ftypeOfInterest}{An input specifying what failure types to compute estimates of incidence for. 
The default value is \code{"all"}, which computes estimates for values \code{unique(ftype)}. Can alternatively
be set to a vector of values found in \code{ftype}.}

\item{trtOfInterest}{An input specifying which levels of \code{trt} are of interest. The default value
is \code{"all"}, which computes estimates for values \code{unique(trt)}. Can alternatively be set to a 
vector of values found in \code{trt}.}

\item{method}{A character specification of how the targeted minimum loss-based estimators should be 
computer, either \code{"mean"} or \code{"hazard"}. The \code{"mean"} specification uses a closed-form
targeted minimum loss-based estimation based on the G-computation formula of Bang and Robins (2005).
The \code{"hazard"} specification uses an iteratively algorithm based on cause-specific hazard functions.
The latter specification has no guarantee of convergence in finite samples. The convergence can be 
influenced by the stopping criteria specified in the \code{tol}. Future versions may implement a closed 
form version of this hazard-based estimator.}

\item{bounds}{A list of bounds... XXX NEED MORE DESCRIPTION HERE XXX}

\item{verbose}{A boolean indicating whether the function should print messages to indicate progress.}

\item{tol}{The stopping criteria when \code{method="hazard"}. The algorithm will continue performing
targeting updates to the initial estimators until the empirical mean of the efficient influence function
is smaller than \code{tol}. The default (\code{1/length(ftime)}) is a sensible value. Larger values can
be used in situations where convergence of the algorithm is an issue; however, this may result in 
large finite-sample bias.}

\item{maxIter}{A maximum number of iterations for the algorithm when \code{method="hazard"}. The 
algorithm will iterate until either the empirical mean of the efficient influence function
is smaller than \code{tol} or until \code{maxIter} iterations have been completed.}
}
\value{
An object of class \code{cv_survtmle}.
\describe{
\item{call}{The call to \code{cv_survtmle}.}
\item{est}{A numeric vector of point estimates -- one for each combination of \code{ftypeOfInterest}
and \code{trtOfInterest}.}
\item{var}{A covariance matrix of all the point estimates.}
\item{meanIC}{The empirical mean of the efficient influence function at the estimated, fluctuated
nuisance parameters. If all goes well, each value should be small. If
\code{method="hazard"} then each value should be less than \code{tol}. If code{method="mean"}
then should be on the order of \code{1e-7}.}
\item{ic}{The efficient influence function at the estimated, fluctuated nuisance parameters,
evaluated on each of the observations (summed over all times). These may be used to perform
post-hoc comparisons.}
\item{ftimeMod}{If \code{returnModels=TRUE} the fit object(s) for the call to \code{glm} or 
\code{SuperLearner} for the outcome regression models. If \code{method="mean"} this will be a list
of length \code{length(ftypeOfInterest)} each of length \code{t0} (one regression for each
failure type and for each time point). If \code{method="hazard"} this will be a list of length
\code{length(ftypeOfInterest)} with one model corresponding with the (pseudo-)hazard for each
cause of failure. If \code{returnModels=FALSE}, this will equal \code{NULL}.}
\item{ctimeMod}{If \code{returnModels=TRUE} the fit object for the call to \code{glm} or 
\code{SuperLearner} for the censoring hazard regression model.  
If \code{returnModels=FALSE}, this will equal \code{NULL}.}
\item{trtMod}{If \code{returnModels=TRUE} the fit object for the call to \code{glm} or 
\code{SuperLearner} for the conditional probability of \code{trt} regression model. 
If \code{returnModels=FALSE}, this will equal \code{NULL}.}
\item{t0}{The timepoint the function was evaluated at.}
}
}
\description{
This function estimates the marginal cumulative incidence for failures of
specified types using cross-validated targeted minimum loss-based estimation. More description to come...
}
