% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmWeights_estimate.R
\name{estimateMSMWeights}
\alias{estimateMSMWeights}
\title{Determine weights for MSM}
\usage{
estimateMSMWeights(
  dat,
  dataList = NULL,
  ntrt,
  uniqtrt,
  adjustVars,
  msm.formula,
  msm.weights,
  returnModels,
  t0 = NULL
)
}
\arguments{
\item{...}{}
}
\value{
A list
}
\description{
TO DO: Add more documentation here.
}
