% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_estimate.R
\name{estimateTreatment}
\alias{estimateTreatment}
\title{Estimate Treatment Mechanisms}
\usage{
estimateTreatment(
  dat,
  adjustVars,
  glm.trt = NULL,
  SL.trt = NULL,
  cvControl,
  returnModels = FALSE,
  verbose = FALSE,
  gtol = 0.001,
  ...
)
}
\arguments{
\item{dat}{An object of class \code{data.frame}. Must have named column
\code{trt}.}

\item{adjustVars}{An object of class \code{data.frame} that will be used
either as the \code{data} argument in a call to \code{\link[stats]{glm}} or
as \code{X} in a call to \code{\link[SuperLearner]{SuperLearner}}.}

\item{glm.trt}{A character formula for the right-hand side of the
\code{\link[stats]{formula}} in a call to \code{\link[stats]{glm}}. See the
documentation of \code{\link{survtmle}} for information. Alternatively,
this could be an object of class \code{\link[stats]{glm}} (as in calls to
this function via \code{\link{timepoints}}), in which case predictions are
obtained using this object with no new fitting.}

\item{SL.trt}{A specification of the \code{SL.library} option of a call to
\code{\link[SuperLearner]{SuperLearner}}. See the documentation of
\code{\link{survtmle}} for more information. Alternatively, this could be
an object of class \code{SuperLearner} (as in calls to this function via
\code{\link{timepoints}}), in which case predictions are obtained using
this object with no new fitting.}

\item{cvControl}{A \code{list} providing control options to be fed directly
into calls to \code{\link[SuperLearner]{SuperLearner}}. This should match
the contents of \code{SuperLearner.CV.control} exactly. For details,
consult the documentation of the \pkg{SuperLearner} package. This is passed
in from \code{\link{mean_tmle}} or \code{\link{hazard_tmle}} via
\code{\link{survtmle}}.}

\item{returnModels}{A \code{logical} indicating whether fitted model objects
should be returned.}

\item{verbose}{A \code{logical} passed to the \code{verbose} option of the
call to \code{\link[SuperLearner]{SuperLearner}}.}

\item{gtol}{The truncation level of predicted trt probabilities to handle
positivity violations.}

\item{...}{Other arguments. Not currently used}
}
\value{
dat The input \code{data.frame} object with two added columns
 corresponding with the conditional probability (given \code{adjustVars}) of
 \code{trt==max(trt)} and \code{trt==min(trt)}.

trtMod If \code{returnModels = TRUE}, the fitted \code{glm} or
 \code{SuperLearner} object. Otherwise, \code{NULL}
}
\description{
This function computes the conditional probability of having
 \code{trt} for each specified level either using \code{\link[stats]{glm}}
 or \code{\link[SuperLearner]{SuperLearner}}. Currently, only two unique
 values of treatment are acceptable. By default the function will compute
 estimates of the conditional probability of \code{trt == max(trt)} and
 compute the probability of \code{trt == min(trt)} as one minus this
 probability.
}
