% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardmsm_estimate.R
\name{getHazardMSMEstimate}
\alias{getHazardMSMEstimate}
\title{Compute the MSM projection parameter}
\usage{
getHazardMSMEstimate(
  dat,
  dataList,
  msm.formula,
  msm.family,
  msmWeightList,
  adjustVarNames,
  ofInterestJ,
  ...
)
}
\arguments{
\item{msm.formula}{}

\item{...}{}
}
\description{
Compute the MSM projection parameter
}
