% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVE.R
\name{getVE}
\alias{getVE}
\title{getVE}
\usage{
getVE(out, tol = 1e-10)
}
\arguments{
\item{out}{A \code{survtmle} object with two types of failure
and two treatment types.}

\item{tol}{Level of incidence below which not to compute VE.
If incidence is very small in the placebo arm the asymmetric scale of VE can
cause very small values of VE.}
}
\value{
A list of data.frames whose first row is the vaccine efficacy against
type 1 endpoints, second row is vaccine efficacy against type 2 endpoints
and third row is vaccine sieve effect. In each row the first entry corresponds
to the effect estimate, the second and third to the lower and upper confidence
limits, and the fourth to the p-value for the Wald test of the
relevant hypothesis test (VE = 0 or VSE = 1).
}
\description{
A function that uses output from a survtmle object to compute
vaccine efficacy and vaccine sieve effect estimates
}

