% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.offset.R
\name{grad.offset}
\alias{grad.offset}
\title{grad.offset}
\usage{
grad.offset(beta, Y, H, offset = NULL)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome}

\item{X}{The design matrix}
}
\value{
Numeric vector of the gradient of the parameter vector
}
\description{
A function that computes the gradient of the for a logistic regression
model with an offset term. Used by 
\code{optim} on occasion.
}

