% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{grad}
\alias{grad}
\title{Gradient for Logistic Regression}
\usage{
grad(beta, Y, X, wts = rep(1, length(Y)))
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome}

\item{X}{The design matrix}

\item{wts}{A \code{numeric} vector of observation-level weights.}
}
\value{
Numeric vector of the gradient of the parameter vector.
}
\description{
A function that computes the gradient of the for a logistic
 regression model. Used by \code{\link[stats]{optim}} on occasion.
}
