% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.offset.R
\name{LogLikelihood.offset}
\alias{LogLikelihood.offset}
\title{LogLikelihood.offset}
\usage{
LogLikelihood.offset(beta, Y, H, offset)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome}

\item{X}{The design matrix}
}
\value{
Numeric of the summed negative log-likelihood loss over observations.
}
\description{
A function that computes the log-likelihood for a logistic regression
model with an offset. Used by 
\code{optim} on occasion.
}

