% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.R
\name{LogLikelihood_offset}
\alias{LogLikelihood_offset}
\title{Log-Likelihood Offset}
\usage{
LogLikelihood_offset(beta, Y, H, offset, wts = rep(1, length(Y)))
}
\arguments{
\item{beta}{A vector of coefficients in a logistic GLM.}

\item{Y}{A \code{numeric} vector of the outcome.}

\item{H}{The \code{matrix} of covariates.}

\item{offset}{The \code{numeric} vector of offsets.}

\item{wts}{A \code{numeric} vector of observation-level weights.}
}
\value{
Numeric of the summed negative log-likelihood loss over
 observations.
}
\description{
Computes the log-likelihood for a logistic regression model
 with an offset. Used by \code{\link[stats]{optim}} on occasion.
}
