% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWideDataListT.R
\name{makeWideDataListT}
\alias{makeWideDataListT}
\title{Convert Long Form Data to List of Wide Form Data (TVC verison)}
\usage{
makeWideDataListT(
  dat,
  allJ,
  uniqtrt,
  trt,
  trtOfInterest,
  adjustVars,
  wideData,
  trtofTime,
  msm.formula = NULL,
  msmWeightList = NULL,
  t0,
  ...
)
}
\arguments{
\item{dat}{The short form \code{data.frame}}

\item{allJ}{Numeric vector indicating the labels of all causes of failure.}

\item{uniqtrt}{The values of \code{trtOfInterest} passed to \code{mean_tmle}.}

\item{adjustVars}{A data.frame of adjustment variables that will be used in
estimating the conditional treatment, censoring, and failure (hazard
or conditional mean) probabilities.}

\item{msm.formula}{A valid right-hand-side of a formula that can include
variables \code{trt} and \code{colnames(adjustVars)}}

\item{msmWeightList}{A list of weights in same format as this list will be (
i.e., first entry corresponding to observed trt, latter to set values
of trt)}

\item{t0}{The timepoint at which \code{survtmle} was called to evaluate.}

\item{...}{Other arguments. Not currently used.}

\item{dataList}{A list of long format \code{data.frame} objects. See
\code{?makeDataList} for more details on formatting.}
}
\value{
A list of \code{data.frame} objects as described above.
}
\description{
The function takes a \code{data.frame} and \code{list} consisting of short
and long format right-censored failure times. The function reshapes the long
format into the wide format needed for calls to \code{mean_tmle}. The list
returned by the function will have number of entries equal to
\code{length(trtOfInterest) + 1}. The first will contain the observed
\code{trt} columns and will set \code{C.t} (the censoring counting process)
equal to the observed value of censoring. The subsequent entries will set
\code{trt} equal to each level of \code{trtOfInterest} and set \code{C.t} to
zero for everyone.
}
