% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.tp.survtmle}
\alias{plot.tp.survtmle}
\title{plot.tp.survtmle}
\usage{
\method{plot}{tp.survtmle}(x, ..., t0, type = c("smooth", "raw"))
}
\arguments{
\item{x}{object of class \code{tp.survtmle} as produced by a sequence of
appropriate calls to \code{survtmle} and \code{timepoints}}

\item{...}{additional arguments passed \code{plot} as necessary}

\item{t0}{numeric indicating the last time point until which estimates are
desired, as originally passed in a call to \code{survtmle}. NOTE: this
is a workaround and should be deprecated soon, when \code{survtmle}
and \code{tp.survtmle} classes are re-written to store t0.}

\item{type}{character describing whether to provide a plot of raw or
smoothened point estimates, with the latter being computed by a call
to \code{stats::isoreg}}
}
\value{
object of class \code{ggplot} containing a lines plot of the raw or
        smoothened point estimates of cumulative incidence across a series of
        time points of interest.
}
\description{
Plot results of cumulative incidence estimates from survtmle
}
\details{
Plotting for both raw and smoothened estimates, the latter by isotonic
regression of the raw point estimates, of cumulative incidence.
}
\examples{

}
