% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.survtmle}
\alias{print.survtmle}
\title{print.survtmle}
\usage{
\method{print}{survtmle}(x, ...)
}
\arguments{
\item{x}{An object of class \code{survtmle}}

\item{...}{Other options (not currently used)}
}
\value{
Prints the estimates of cumulative incidence and the diagonal
        of the estimated covariance matrix.
}
\description{
The print method for an object of class \code{survtmle}
}
