% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.tp.survtmle}
\alias{print.tp.survtmle}
\title{print.tp.survtmle}
\usage{
\method{print}{tp.survtmle}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tp.survtmle}.}

\item{...}{Other options (not currently used).}
}
\value{
Prints the estimates of cumulative incidence and the diagonal
        of the estimated covariance matrix.
}
\description{
The print method for a timepoints object of class \code{tp.survtmle}
}
