% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglm.R
\name{fast_glm}
\alias{fast_glm}
\title{Faster Yet Robust Generalized Linear Models}
\usage{
fast_glm(reg_form, data, err_fam, start = NULL)
}
\arguments{
\item{reg_form}{...}

\item{data}{...}

\item{err_fam}{...}

\item{start}{...}
}
\value{
...
}
\description{
words, words, words, ...
}
