% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglm.R
\name{fast_glm}
\alias{fast_glm}
\title{Faster Yet Robust Generalized Linear Models}
\usage{
fast_glm(reg_form, data, family, flavor = "fast", start = NULL)
}
\arguments{
\item{reg_form}{...}

\item{data}{...}

\item{family}{...}

\item{flavor}{Character: do you want to live dangerously?}

\item{start}{...}
}
\value{
...
}
\description{
words, words, words, ...
}
