% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fast.R
\name{fast_glm}
\alias{fast_glm}
\title{Wrapper for faster Generalized Linear Models}
\usage{
fast_glm(reg_form, data, family, ...)
}
\arguments{
\item{reg_form}{Object of class \code{formula} indicating the regression to
be fit.}

\item{data}{Object of class \code{data.frame} containing the data.}

\item{family}{Object of class \code{family} from package \code{stats}
indicating the error distribution. Appropriate options are limited to
\code{gaussian} and \code{binomial}.}

\item{...}{Additional arguments passed to \code{glm} or \code{speedglm}.}
}
\value{
Object of class \code{glm} or \code{speedglm}.
}
\description{
A convenience utility to fit regression models more quickly in the main
internal functions for estimation, which usually require logistic regression.
Use of \code{speedglm} appears to provide roughly an order of magnitude
improvement in speed when compared to \code{glm} in custom benchmarks.
}
