% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format.perc}
\alias{format.perc}
\title{format.perc}
\usage{
\method{format}{perc}(probs, digits)
}
\arguments{
\item{probs}{Probabilities}

\item{digits}{Number of digits to round to}
}
\description{
Copied from package \code{stats}.
}
