% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{grad}
\alias{grad}
\title{Gradient for Logistic Regression}
\usage{
grad(beta, Y, X)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome}

\item{X}{The design matrix}
}
\value{
Numeric vector of the gradient of the parameter vector
}
\description{
A function that computes the gradient of the for a logistic regression
model. Used by \code{optim} on occasion.
}
