% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{grad_offset}
\alias{grad_offset}
\title{Gradient for Logistic Regression with Offsets}
\usage{
grad_offset(beta, Y, H, offset = NULL)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome}

\item{H}{The covariate matrix}

\item{offset}{The offset vector}
}
\value{
Numeric vector of the gradient of the parameter vector
}
\description{
A function that computes the gradient of the for a logistic regression
model with an offset term. Used by \code{optim} on occasion.
}
