% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Log-Likelihood}
\usage{
LogLikelihood(beta, X, Y, wts = rep(1, length(Y)))
}
\arguments{
\item{beta}{A \code{numeric} vector of coefficients in a logistic GLM.}

\item{X}{The design matrix.}

\item{Y}{The outcome.}

\item{wts}{A \code{numeric} vector of observation-level weights.}
}
\value{
Numeric of the summed negative log-likelihood loss over
 observations.
}
\description{
Computes the log-likelihood for a model. Used by
 \code{\link[stats]{optim}} on occasion.
}
