% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Log-Likelihood}
\usage{
LogLikelihood(beta, X, Y)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic GLM.}

\item{X}{The design matrix.}

\item{Y}{The outcome.}
}
\value{
Numeric of the summed negative log-likelihood loss over observations.
}
\description{
Computes the log-likelihood for a model. Used by \code{optim} on occasion.
}
