% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.offset.R
\name{LogLikelihood_offset}
\alias{LogLikelihood_offset}
\title{LogLikelihood_offset}
\usage{
LogLikelihood_offset(beta, Y, H, offset)
}
\arguments{
\item{beta}{A vector of coefficients in a logistic glm}

\item{Y}{The outcome vector}

\item{H}{The covariate matrix}

\item{offset}{The offset vector}
}
\value{
Numeric of the summed negative log-likelihood loss over observations.
}
\description{
A function that computes the log-likelihood for a logistic regression
model with an offset. Used by 
\code{optim} on occasion.
}
