% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.tp.survtmle}
\alias{plot.tp.survtmle}
\title{Plot results of cumulative incidence estimates from survtmle}
\usage{
\method{plot}{tp.survtmle}(x, ..., type = c("smooth", "raw"))
}
\arguments{
\item{x}{object of class \code{tp.survtmle} as produced by a sequence of
appropriate calls to \code{survtmle} and \code{timepoints}}

\item{...}{additional arguments passed \code{plot} as necessary}

\item{type}{character describing whether to provide a plot of raw or
smoothened point estimates, with the latter being computed by a call
to \code{stats::isoreg}}
}
\value{
object of class \code{ggplot} containing a lines plot of the raw or
        smoothened point estimates of cumulative incidence across a series of
        time points of interest.
}
\description{
Plotting for both raw and smoothened estimates, the latter by isotonic
regression of the raw point estimates, of cumulative incidence.
}
